requirements.txt
pbr>=0.6,!=0.7,<1.0
python-keystoneclient>=1.2.0,<1.4.0
python-neutronclient>=2.4.0,<2.5.0
python-novaclient>=2.22.0,<2.24.0
python-openstackclient>=1.0.3,<1.1.0
PyYAML>=3.1.0
oslo.config>=1.9.3,<1.10.0  # Apache-2.0
oslo.i18n>=1.5.0,<1.6.0  # Apache-2.0
oslo.log>=1.0.0,<1.1.0  # Apache-2.0